[python3] Reset to version 3.10.7 in ports
This is for vcpkg operating in _classic_ (a.k.a. command-line) mode.
In this mode, vcpkg is not aware of versions and simply installs
whatever is in ports/ subdir.

Currently we're using classic mode in vcpkg-privates CI, and can't build
Python 3.11, hence this commit. It was created by this sequence of
commands (a0f1... SHA is taken from versions/p-/python3/vcpkg.json file):

       mkdir bak
       git read-tree --prefix bak -u a0f1ead9323b9521f3c3016469f028be712c0252
       rm -rf ports/python3/
       mv x ports/python3/
       git add -A
       git commit